+++ src/index.js
@@ -1,6 +1,8 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 
-import App from './app';
 
-ReactDOM.render(<App/> , document.getElementById('root'));
+import App from './App.js';
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(<App />);