+++ Changelog.md
@@ -1,3 +1,11 @@
+CL-0.0.02: Making API call.
+- Updated reactDom render method as it was giving error in console due to v18 of react.
+- Added api call.
+- Added search svg.
+- Added whole project css.
+- Added MovieCard component.
+- Used useState hooks to store data,state and used useEffect to make API call at start/loading of App component.
+
 CL-0.0.01: Bootstrapping.
  - Created the React app using CRA template.
  - Added secret keys in env variables.